+++ database/migrations/2021_04_16_153205_create_product_brands_table.php
@@ -15,7 +15,7 @@ class CreateProductBrandsTable extends Migration
     {
         Schema::create('product_brands', function (Blueprint $table) {
             $table->id();
-            $table->integer('privacy')->default(1);
+            $table->smallInteger('privacy')->default(1);
             $table->string('title');
             $table->string('url');
             $table->text('overview')->nullable();
